requirements.txt
langchain-chroma>=0.1.2
langchain_huggingface
PyPDF2
streamlit
torch==2.2.1
langchain_community
vswarm @ git+https://github.com/Vu0401/vswarm.git